# WIP: scale out TileLoop, to speed up bigger maps