## Motivation / Problem

I was bored, and wanted to do something fun.

This idea was living in my head for a while now, and when JGRennison showed me he already implemented a worker pool, it just had to be tried.

Mostly I was curious if the TileLoop can be made faster by using a worker-pool. The answer is a very clear and loud: yes.
The next, much harder, assignment will be: make all TileLoops thread-safe. They aren't. So yeah .. that is fun.
But if we go this route, it does speed up the game, for those who have more than 1 core (read: according to Steam survey, 99% of our users), especially on bigger maps.

This was btw the trigger to look into this: Steam says almost nobody has a single-core CPU anymore. So we don't have to worry that much anymore what the performance loss is on those machines; only what the gain is on multi-core CPUs :) Which is a good thing!

## Description

This is a proof-of-concept, to show that something like this will help, but requires some work on the current TileLoop implementations.

On a 4kx4k map the TileLoop goes from 17ms to 3ms on my machine, and a 4kx4k map finally becomes playable (in Debug-mode).
For Release-mode it is from 5ms to 1.5ms.

Currently it queues 1024 tile-updates on a single worker, in an attempt to balance how much we push in a std::queue and how long a worker takes to execute the job. This number is not tuned, but just a number that looked nice.

With this change, it means a TileLoop should only deal with its local Tile, and nothing else. 100% sure this currently isn't always the case, so expect issues when using this.

The main one that kept crashing was the updating of the towns coordinates. But the problem below that is of course that multiple threads updated how big the town is (as that updates the coordinates). So the current fix is .. not a fix, but a workaround.

`worker_thread` is a copy from https://github.com/JGRennison/OpenTTD-patches/blob/b776dcbd173fd09bfa1c9904a0606d4709cdb49b/src/worker_thread.cpp, with small changes (changed the function signature and there is always a worker that will be notified when something got inserted in the queue). Of course this is a dirty solution, and we should go for a more generic approach, but again: proof-of-concept. Tnx JGRennison for the work here :)

Owh, and of course: a picture, or it didn't happen.

Badly tuned batch-size:
![image](https://user-images.githubusercontent.com/1663690/230742347-b5af48e1-fb83-4f54-b2ec-6524b091f07f.png)

Better tuned batch-size:
![image](https://user-images.githubusercontent.com/1663690/230743634-bae5fb4b-a8c7-4ef0-a681-562fb1a8e8e7.png)

Als, as it turns out, release-builds handle a batch of 256 so quickly, that the workers had hardly anything to do. This will be one of the challenges to get such PR just right. As the batch-size will differ per CPU.
- Going too big? There is no scaling out.
- Going too small? The overhead is higher than not threading.
- There is a sweet-spot in between.

## Limitations

SO MANY! Where to begin? Just read the description, okay?

## Checklist for review

Some things are not automated, and forgotten often. This list is a reminder for the reviewers.
* The bug fix is important enough to be backported? (label: 'backport requested')
* This PR touches english.txt or translations? Check the [guidelines](https://github.com/OpenTTD/OpenTTD/blob/master/docs/eints.md)
* This PR affects the save game format? (label 'savegame upgrade')
* This PR affects the GS/AI API? (label 'needs review: Script API')
    * ai_changelog.hpp, gs_changelog.hpp need updating.
    * The compatibility wrappers (compat_*.nut) need updating.
* This PR affects the NewGRF API? (label 'needs review: NewGRF')
    * newgrf_debug_data.h may need updating.
    * [PR must be added to API tracker](https://wiki.openttd.org/en/Development/NewGRF/Specification%20Status)