@TheMowgliMan 2023-04-08
I would like to see this merged.

The issue is, previous attempts to add multithreading we're rejected.

On Sat, Apr 8, 2023, 4:55 PM Patric Stout ***@***.***> wrote:

> Motivation / Problem
>
> I was bored, and wanted to do something fun.
>
> This idea was living in my head for a while now, and when JGRennison
> showed me he already implemented a worker pool, it just had to be tried.
>
> Mostly I was curious if the TileLoop can be made faster by using a
> worker-pool. The answer is a very clear and loud: yes.
> The next, much harder, assignment will be: make all TileLoops thread-safe.
> They aren't. So yeah .. that is fun.
> But if we go this route, it does speed up the game, for those who have
> more than 1 core (read: according to Steam survey, 99% of our users),
> especially on bigger maps.
> Description
>
> This is a proof-of-concept, to show that something like this will help,
> but requires some work on the current TileLoop implementations.
>
> On a 4kx4k map the TileLoop goes from 17ms to 3ms on my machine, and a
> 4kx4k map finally becomes playable (in Debug-mode).
>
> Currently it queues 256 tile-updates on a single worker, in an attempt to
> balance how much we push in a std::queue and how long a worker takes to
> execute the job. This number is not tuned, but just a number that looked
> nice.
>
> With this change, it means a TileLoop should only deal with its local
> Tile, and nothing else. 100% sure this currently isn't always the case, so
> expect issues when using this.
>
> The main one that kept crashing was the updating of the towns coordinates.
> But the problem below that is of course that multiple threads updated how
> big the town is (as that updates the coordinates). So the current fix is ..
> not a fix, but a workaround.
> Limitations Checklist for review
>
> Some things are not automated, and forgotten often. This list is a
> reminder for the reviewers.
>
>    - The bug fix is important enough to be backported? (label: 'backport
>    requested')
>    - This PR touches english.txt or translations? Check the guidelines
>    <https://github.com/OpenTTD/OpenTTD/blob/master/docs/eints.md>
>    - This PR affects the save game format? (label 'savegame upgrade')
>    - This PR affects the GS/AI API? (label 'needs review: Script API')
>       - ai_changelog.hpp, gs_changelog.hpp need updating.
>       - The compatibility wrappers (compat_*.nut) need updating.
>    - This PR affects the NewGRF API? (label 'needs review: NewGRF')
>       - newgrf_debug_data.h may need updating.
>       - PR must be added to API tracker
>       <https://wiki.openttd.org/en/Development/NewGRF/Specification%20Status>
>
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/OpenTTD/OpenTTD/pull/10614
> Commit Summary
>
>    - a78790a
>    <https://github.com/OpenTTD/OpenTTD/pull/10614/commits/a78790a01eb3e974759f4585c54136f87b9261d4>
>    WIP: scale out TileLoop, to help with big maps
>
> File Changes
>
> (6 files <https://github.com/OpenTTD/OpenTTD/pull/10614/files>)
>
>    - *M* src/CMakeLists.txt
>    <https://github.com/OpenTTD/OpenTTD/pull/10614/files#diff-148715d6ea0c0ea0a346af3f6bd610d010d490eca35ac6a9b408748f7ca9e3f4>
>    (2)
>    - *M* src/landscape.cpp
>    <https://github.com/OpenTTD/OpenTTD/pull/10614/files#diff-578266c8c75b8543c3d724834f1ed634626651d6a9446c9fa02073a80c67d3ca>
>    (25)
>    - *M* src/openttd.cpp
>    <https://github.com/OpenTTD/OpenTTD/pull/10614/files#diff-84e5bd1673aba7f0fbb76c27ae19a02693636d290501c5af53d144ef05a88c08>
>    (3)
>    - *M* src/town_cmd.cpp
>    <https://github.com/OpenTTD/OpenTTD/pull/10614/files#diff-88e5b6c0a7c202f09d40e5a721b54aeec678086c7116459241d4fab02f30a33c>
>    (6)
>    - *A* src/worker_thread.cpp
>    <https://github.com/OpenTTD/OpenTTD/pull/10614/files#diff-2eb0c2685a1fab75a0b1cf01d05a96641470da15356efab8d0750d4f8108088a>
>    (90)
>    - *A* src/worker_thread.h
>    <https://github.com/OpenTTD/OpenTTD/pull/10614/files#diff-e8ec427b9a84a91af2db6e21ddd9fbe12275cb25795bdf3db372c426da7e71c3>
>    (59)
>
> Patch Links:
>
>    - https://github.com/OpenTTD/OpenTTD/pull/10614.patch
>    - https://github.com/OpenTTD/OpenTTD/pull/10614.diff
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/OpenTTD/OpenTTD/pull/10614>, or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AW7ZWTKXLB2DGBE5DHYW4GTXAHGENANCNFSM6AAAAAAWXVJNUQ>
> .
> You are receiving this because you are subscribed to this thread.Message
> ID: ***@***.***>
>