@TrueBrain 2023-04-08
And I would like a mansion! But in case the description wasn't clear enough, but I hope it was: this is YEARS away from being finished. Not days, not weeks, not even months. This is so far WIP, that WIP needed to redefine itself to include this amount of WIP.

So don't get your hopes up.

And no, "previous attempts" weren't rejected. They didn't work. Which is something completely different. Very similar to this PR: it doesn't actually work.