@TrueBrain 2023-04-08
CLang Thread Sanitizer reports at least 20 places where there is a "data race" (where two threads are accessing the same memory). 

Some are easy (dirty-blocks) and can be solved with a lock, others are .. interesting.

Like, `TileLoop_Trees` can plant trees on nearby tiles. This .. is problematic. Possibly we can resolve that by making an "action-to-take" queue, where the TileLoop can insert functions to execute when all TileLoops are done.

Yeah .. work to be done here :)