@JGRennison 2023-04-09
The smoke plumes above animated vanilla industries are visually broken, and they eventually triggered a crash.
These use effect vehicles, so create entries in the global vehicle pool.

Industries in general can trigger all sort of stuff which reads and writes non tile local state from their tile loop handlers.
Possibly the whole industry tile loop handler should be queued.