@ldpl 2023-04-09
In theory it can be solved by having two tile loop modes, concurrent and general. In concurrent mode tile can request a general run that is not parallelized. Though not sure what is the ratio of tiles that can be run fully concurrent and so if it's worth the overhead.